Move utils
Move utils

#4053
